Done some robustness for updating data. Also got date to work in updating data. Slight UI changes. Date still doesnt work with the database properly

+++ app/src/main/java/com/example/jakeshepherd/moneymanager/DeleteData.java
@@ -1,5 +1,6 @@
 package com.example.jakeshepherd.moneymanager;
 
+import android.content.Context;
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
@@ -7,6 +8,7 @@ import android.util.DisplayMetrics;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
+import android.widget.Toast;
 
 import com.example.jakeshepherd.moneymanager.Database;
 import com.example.jakeshepherd.moneymanager.R;
@@ -16,10 +18,6 @@ public class DeleteData extends AppCompatActivity {
 
     Database db;
 
-    /**
-     * TODO add robustness
-     * @param savedInstanceState
-     */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -49,9 +47,24 @@ public class DeleteData extends AppCompatActivity {
                  * this will delete the specified row
                  * could look into how this is implemented in order to search for rows.
                  */
-                db.deleteRowData(billNum);
-                Intent viewBills = new Intent(getBaseContext(), ViewBills.class);
-                startActivity(viewBills);
+
+                if(billNum.equals("")){
+                    billToDelete.setError("Fill in");
+                    showMessage("Please enter a bill number");
+                }else{
+                    int test = db.deleteRowData(billNum);
+
+                    if(test == 0){
+                        showMessage("This bill number does not exist, please try again.");
+                    }else{
+                        showMessage("Bill number " + billNum + " has been deleted");
+                        Intent viewBills = new Intent(getBaseContext(), ViewBills.class);
+                        startActivity(viewBills);
+                    }
+                }
+
+
+
             }
         });
 
@@ -62,5 +75,11 @@ public class DeleteData extends AppCompatActivity {
                 startActivity(back);
             }
         });
+    }
+
+    private void showMessage(String msg) {
+        Context c = getApplicationContext();
+        Toast t = Toast.makeText(c, msg, Toast.LENGTH_SHORT);
+        t.show();
     }
 }
